+++ lib/common/http.dart
@@ -1,15 +1,20 @@
 import 'dart:io';
 
 import 'package:fl_clash/common/common.dart';
-import 'package:fl_clash/controller.dart';
+import 'package:fl_clash/providers/providers.dart';
+import 'package:fl_clash/state.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 class FlClashHttpOverrides extends HttpOverrides {
   static String handleFindProxy(Uri url) {
     if ([localhost].contains(url.host)) {
       return 'DIRECT';
     }
-    final port = appController.config.patchClashConfig.mixedPort;
-    final isStart = appController.isStart;
+    final ref = globalState.container;
+    final port = ref.read(
+      patchClashConfigProvider.select((state) => state.port),
+    );
+    final isStart = ref.read(isStartProvider);
     commonPrint.log('find $url proxy:$isStart');
     if (!isStart) return 'DIRECT';
     return 'PROXY localhost:$port';
