+++ lib/common/print.dart
@@ -1,6 +1,8 @@
 import 'package:fl_clash/controller.dart';
 import 'package:fl_clash/enum/enum.dart';
 import 'package:fl_clash/models/models.dart';
+import 'package:fl_clash/providers/app.dart';
+import 'package:fl_clash/state.dart';
 import 'package:flutter/material.dart';
 
 class CommonPrint {
@@ -19,7 +21,9 @@ class CommonPrint {
     if (!appController.isAttach) {
       return;
     }
-    appController.addLog(Log.app(payload).copyWith(logLevel: logLevel));
+    globalState.container
+        .read(logsProvider.notifier)
+        .add(Log.app(payload).copyWith(logLevel: logLevel));
   }
 }
