+++ lib/common/tray.dart
@@ -1,8 +1,9 @@
 import 'dart:io';
 
-import 'package:fl_clash/controller.dart';
 import 'package:fl_clash/enum/enum.dart';
 import 'package:fl_clash/models/models.dart';
+import 'package:fl_clash/providers/providers.dart';
+import 'package:fl_clash/state.dart';
 import 'package:flutter/services.dart';
 import 'package:intl/intl.dart';
 import 'package:tray_manager/tray_manager.dart';
@@ -70,6 +71,10 @@ class Tray {
       );
     }
     List<MenuItem> menuItems = [];
+    final ref = globalState.container;
+    final commonAction = ref.read(commonActionProvider.notifier);
+    final systemAction = ref.read(systemActionProvider.notifier);
+    final setupAction = ref.read(setupActionProvider.notifier);
     final appLocalizations = currentAppLocalizations;
     final showMenuItem = MenuItem(
       label: appLocalizations.show,
@@ -81,7 +86,7 @@ class Tray {
     final startMenuItem = MenuItem.checkbox(
       label: trayState.isStart ? appLocalizations.stop : appLocalizations.start,
       onClick: (_) async {
-        appController.updateStart();
+        commonAction.updateStart();
       },
       checked: false,
     );
@@ -90,7 +95,7 @@ class Tray {
       final speedStatistics = MenuItem.checkbox(
         label: appLocalizations.speedStatistics,
         onClick: (_) async {
-          appController.updateSpeedStatistics();
+          commonAction.updateSpeedStatistics();
         },
         checked: trayState.showTrayTitle,
       );
@@ -102,7 +107,7 @@ class Tray {
         MenuItem.checkbox(
           label: Intl.message(mode.name),
           onClick: (_) {
-            appController.changeMode(mode);
+            setupAction.changeMode(mode);
           },
           checked: mode == trayState.mode,
         ),
@@ -117,13 +122,14 @@ class Tray {
             MenuItem.checkbox(
               label: proxy.name,
               checked:
-                  appController.getSelectedProxyName(group.name) == proxy.name,
+                  ref.read(selectedProxyNameProvider(group.name)) == proxy.name,
               onClick: (_) {
-                appController.updateCurrentSelectedMap(group.name, proxy.name);
-                appController.changeProxy(
-                  groupName: group.name,
-                  proxyName: proxy.name,
-                );
+                ref
+                    .read(profilesActionProvider.notifier)
+                    .updateCurrentSelectedMap(group.name, proxy.name);
+                ref
+                    .read(proxiesActionProvider.notifier)
+                    .changeProxy(groupName: group.name, proxyName: proxy.name);
               },
             ),
           );
@@ -144,7 +150,7 @@ class Tray {
         MenuItem.checkbox(
           label: appLocalizations.tun,
           onClick: (_) {
-            appController.updateTun();
+            systemAction.updateTun();
           },
           checked: trayState.tunEnable,
         ),
@@ -153,7 +159,7 @@ class Tray {
         MenuItem.checkbox(
           label: appLocalizations.systemProxy,
           onClick: (_) {
-            appController.updateSystemProxy();
+            systemAction.updateSystemProxy();
           },
           checked: trayState.systemProxy,
         ),
@@ -163,7 +169,7 @@ class Tray {
     final autoStartMenuItem = MenuItem.checkbox(
       label: appLocalizations.autoLaunch,
       onClick: (_) async {
-        appController.updateAutoLaunch();
+        systemAction.updateAutoLaunch();
       },
       checked: trayState.autoLaunch,
     );
@@ -179,7 +185,7 @@ class Tray {
     final exitMenuItem = MenuItem(
       label: appLocalizations.exit,
       onClick: (_) async {
-        await appController.handleExit();
+        await systemAction.handleExit();
       },
     );
     menuItems.add(exitMenuItem);
