+++ lib/core/lib.dart
@@ -1,10 +1,11 @@
 import 'dart:async';
 
 import 'package:fl_clash/common/common.dart';
-import 'package:fl_clash/controller.dart';
 import 'package:fl_clash/enum/enum.dart';
 import 'package:fl_clash/models/core.dart';
 import 'package:fl_clash/plugins/service.dart';
+import 'package:fl_clash/providers/providers.dart';
+import 'package:fl_clash/state.dart';
 
 import 'interface.dart';
 
@@ -22,7 +23,9 @@ class CoreLib extends CoreHandlerInterface {
       return res ?? '';
     }
     _connectedCompleter.complete(true);
-    final syncRes = await service?.syncState(appController.sharedState);
+    final syncRes = await service?.syncState(
+      globalState.container.read(sharedStateProvider),
+    );
     return syncRes ?? '';
   }
