+++ lib/manager/app_manager.dart
@@ -1,7 +1,6 @@
 import 'dart:async';
 
 import 'package:fl_clash/common/common.dart';
-import 'package:fl_clash/controller.dart';
 import 'package:fl_clash/enum/enum.dart';
 import 'package:fl_clash/manager/window_manager.dart';
 import 'package:fl_clash/providers/providers.dart';
@@ -33,12 +32,16 @@ class _AppStateManagerState extends ConsumerState<AppStateManager>
     });
     ref.listenManual(configProvider, (prev, next) {
       if (prev != next) {
-        appController.savePreferencesDebounce();
+        globalState.container
+            .read(storeActionProvider.notifier)
+            .savePreferencesDebounce();
       }
     });
     ref.listenManual(needUpdateGroupsProvider, (prev, next) {
       if (prev != next) {
-        appController.updateGroupsDebounce();
+        globalState.container
+            .read(proxiesActionProvider.notifier)
+            .updateGroupsDebounce();
       }
     });
     if (window == null) {
@@ -68,9 +71,10 @@ class _AppStateManagerState extends ConsumerState<AppStateManager>
     if (state == AppLifecycleState.resumed) {
       render?.resume();
       WidgetsBinding.instance.addPostFrameCallback((_) {
-        appController.tryCheckIp();
+        final ref = globalState.container;
+        ref.read(setupActionProvider.notifier).tryCheckIp();
         if (system.isAndroid) {
-          appController.tryStartCore();
+          ref.read(coreActionProvider.notifier).tryStartCore();
         }
       });
     }
@@ -78,7 +82,7 @@ class _AppStateManagerState extends ConsumerState<AppStateManager>
 
   @override
   void didChangePlatformBrightness() {
-    appController.updateBrightness();
+    globalState.container.read(themeActionProvider.notifier).updateBrightness();
   }
 
   @override
@@ -164,6 +168,12 @@ class AppSidebarContainer extends ConsumerWidget {
     });
   }
 
+  void _handleToPage(PageLabel pageLabel) {
+    globalState.container
+        .read(currentPageLabelProvider.notifier)
+        .toPage(pageLabel);
+  }
+
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final navigationState = ref.watch(navigationStateProvider);
@@ -216,9 +226,7 @@ class AppSidebarContainer extends ConsumerWidget {
                                 )
                                 .toList(),
                             onDestinationSelected: (index) {
-                              appController.toPage(
-                                navigationItems[index].label,
-                              );
+                              _handleToPage(navigationItems[index].label);
                             },
                             extended: false,
                             selectedIndex: currentIndex,
