+++ lib/manager/core_manager.dart
@@ -1,8 +1,8 @@
 import 'package:fl_clash/common/common.dart';
-import 'package:fl_clash/controller.dart';
 import 'package:fl_clash/core/core.dart';
 import 'package:fl_clash/enum/enum.dart';
 import 'package:fl_clash/models/models.dart';
+import 'package:fl_clash/providers/action.dart';
 import 'package:fl_clash/providers/app.dart';
 import 'package:fl_clash/providers/config.dart';
 import 'package:fl_clash/providers/state.dart';
@@ -34,13 +34,13 @@ class _CoreContainerState extends ConsumerState<CoreManager>
       currentSetupStateProvider.select((state) => state?.profileId),
       (prev, next) {
         if (prev != next) {
-          appController.fullSetup();
+          ref.read(setupActionProvider.notifier).fullSetup();
         }
       },
     );
     ref.listenManual(updateParamsProvider, (prev, next) {
       if (prev != next) {
-        appController.updateConfigDebounce();
+        ref.read(setupActionProvider.notifier).updateConfigDebounce();
       }
     });
     ref.listenManual(appSettingProvider.select((state) => state.openLogs), (
@@ -64,15 +64,16 @@ class _CoreContainerState extends ConsumerState<CoreManager>
   @override
   Future<void> onDelay(Delay delay) async {
     super.onDelay(delay);
-    appController.setDelay(delay);
+    final proxiesAction = ref.read(proxiesActionProvider.notifier);
+    proxiesAction.setDelay(delay);
     debouncer.call(FunctionTag.updateDelay, () async {
-      appController.updateGroupsDebounce();
+      proxiesAction.updateGroupsDebounce();
     }, duration: const Duration(milliseconds: 5000));
   }
 
   @override
   void onLog(Log log) {
-    ref.read(logsProvider.notifier).addLog(log);
+    ref.read(logsProvider.notifier).add(log);
     if (log.logLevel == LogLevel.error) {
       globalState.showNotifier(log.payload);
     }
@@ -87,11 +88,12 @@ class _CoreContainerState extends ConsumerState<CoreManager>
 
   @override
   Future<void> onLoaded(String providerName) async {
+    final ref = globalState.container;
     ref
         .read(providersProvider.notifier)
         .setProvider(await coreController.getExternalProvider(providerName));
     debouncer.call(FunctionTag.loadedProvider, () async {
-      appController.updateGroupsDebounce();
+      ref.read(proxiesActionProvider.notifier).updateGroupsDebounce();
     }, duration: const Duration(milliseconds: 5000));
     super.onLoaded(providerName);
   }
