+++ lib/manager/hotkey_manager.dart
@@ -1,8 +1,9 @@
 import 'package:fl_clash/common/common.dart';
-import 'package:fl_clash/controller.dart';
 import 'package:fl_clash/enum/enum.dart';
 import 'package:fl_clash/models/common.dart';
+import 'package:fl_clash/providers/action.dart';
 import 'package:fl_clash/providers/config.dart';
+import 'package:fl_clash/state.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -29,17 +30,20 @@ class _HotKeyManagerState extends ConsumerState<HotKeyManager> {
   }
 
   Future<void> _handleHotKeyAction(HotAction action) async {
+    final ref = globalState.container;
+    final commonAction = ref.read(commonActionProvider.notifier);
+    final systemAction = ref.read(systemActionProvider.notifier);
     switch (action) {
       case HotAction.mode:
-        appController.updateMode();
+        commonAction.updateMode();
       case HotAction.start:
-        appController.updateStart();
+        commonAction.updateStart();
       case HotAction.view:
-        appController.updateVisible();
+        systemAction.updateVisible();
       case HotAction.proxy:
-        appController.updateSystemProxy();
+        systemAction.updateSystemProxy();
       case HotAction.tun:
-        appController.updateTun();
+        systemAction.updateTun();
     }
   }
 
@@ -78,7 +82,9 @@ class _HotKeyManagerState extends ConsumerState<HotKeyManager> {
       child: Actions(
         actions: {
           CloseWindowIntent: CallbackAction<CloseWindowIntent>(
-            onInvoke: (_) => appController.handleBackOrExit(),
+            onInvoke: (_) => globalState.container
+                .read(systemActionProvider.notifier)
+                .handleBackOrExit(),
           ),
           DoNothingIntent: CallbackAction<DoNothingIntent>(
             onInvoke: (_) => null,
