+++ lib/manager/vpn_manager.dart
@@ -1,7 +1,7 @@
 import 'package:fl_clash/common/common.dart';
-import 'package:fl_clash/controller.dart';
 import 'package:fl_clash/enum/enum.dart';
 import 'package:fl_clash/models/models.dart';
+import 'package:fl_clash/providers/action.dart';
 import 'package:fl_clash/providers/state.dart';
 import 'package:fl_clash/state.dart';
 import 'package:flutter/material.dart';
@@ -40,7 +40,9 @@ class _VpnContainerState extends ConsumerState<VpnManager> {
             actionText: currentAppLocalizations.restart,
             action: () async {
               await globalState.handleStop();
-              await appController.updateStatus(true);
+              await globalState.container
+                  .read(setupActionProvider.notifier)
+                  .updateStatus(true);
             },
           ),
         );
