+++ lib/manager/window_manager.dart
@@ -1,9 +1,9 @@
 import 'dart:async';
 
 import 'package:fl_clash/common/common.dart';
-import 'package:fl_clash/controller.dart';
 import 'package:fl_clash/enum/enum.dart';
 import 'package:fl_clash/providers/providers.dart';
+import 'package:fl_clash/state.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:window_ext/window_ext.dart';
@@ -44,7 +44,7 @@ class _WindowContainerState extends ConsumerState<WindowManager>
 
   @override
   void onWindowClose() async {
-    await appController.handleBackOrExit();
+    await ref.read(systemActionProvider.notifier).handleBackOrExit();
     super.onWindowClose();
   }
 
@@ -57,7 +57,7 @@ class _WindowContainerState extends ConsumerState<WindowManager>
 
   @override
   Future<void> onShouldTerminate() async {
-    await appController.handleExit();
+    await ref.read(systemActionProvider.notifier).handleExit();
     super.onShouldTerminate();
   }
 
@@ -65,8 +65,9 @@ class _WindowContainerState extends ConsumerState<WindowManager>
   void onWindowMoved() {
     super.onWindowMoved();
     windowManager.getPosition().then((offset) {
-      ref.read(windowSettingProvider.notifier);
-      // .update((state) => state.copyWith(top: offset.dy, left: offset.dx));
+      ref
+          .read(windowSettingProvider.notifier)
+          .update((state) => state.copyWith(top: offset.dy, left: offset.dx));
     });
   }
 
@@ -83,7 +84,7 @@ class _WindowContainerState extends ConsumerState<WindowManager>
 
   @override
   void onWindowMinimize() async {
-    appController.savePreferencesDebounce();
+    ref.read(storeActionProvider.notifier).savePreferencesDebounce();
     commonPrint.log('minimize');
     render?.pause();
     super.onWindowMinimize();
@@ -223,7 +224,9 @@ class _WindowHeaderState extends State<WindowHeader> {
         ),
         IconButton(
           onPressed: () {
-            appController.handleBackOrExit();
+            globalState.container
+                .read(systemActionProvider.notifier)
+                .handleBackOrExit();
           },
           icon: const Icon(Icons.close),
         ),
