+++ lib/pages/home.dart
@@ -1,9 +1,9 @@
 import 'package:fl_clash/common/common.dart';
-import 'package:fl_clash/controller.dart';
 import 'package:fl_clash/enum/enum.dart';
 import 'package:fl_clash/manager/app_manager.dart';
 import 'package:fl_clash/models/common.dart';
 import 'package:fl_clash/providers/providers.dart';
+import 'package:fl_clash/state.dart';
 import 'package:fl_clash/widgets/widgets.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
@@ -14,6 +14,12 @@ typedef OnSelected = void Function(int index);
 
 class HomePage extends StatelessWidget {
   const HomePage({super.key});
+
+  void _handleToPage(PageLabel pageLabel) {
+    globalState.container
+        .read(currentPageLabelProvider.notifier)
+        .toPage(pageLabel);
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -42,7 +48,7 @@ class HomePage extends StatelessWidget {
                       )
                       .toList(),
                   onDestinationSelected: (index) {
-                    appController.toPage(navigationItems[index].label);
+                    _handleToPage(navigationItems[index].label);
                   },
                   selectedIndex: currentIndex,
                 ),
@@ -279,7 +285,9 @@ class HomeBackScopeContainer extends ConsumerWidget {
         if (canPop) {
           Navigator.of(realContext).pop();
         } else {
-          await appController.handleBackOrExit();
+          await globalState.container
+              .read(systemActionProvider.notifier)
+              .handleBackOrExit();
         }
         return false;
       },
