+++ lib/providers/providers.dart
@@ -1,3 +1,4 @@
+export 'action.dart';
 export 'app.dart';
 export 'config.dart';
 export 'database.dart';
