+++ lib/views/developer.dart
@@ -1,8 +1,9 @@
 import 'package:fl_clash/common/common.dart';
-import 'package:fl_clash/controller.dart';
 import 'package:fl_clash/core/controller.dart';
 import 'package:fl_clash/enum/enum.dart';
 import 'package:fl_clash/models/common.dart';
+import 'package:fl_clash/providers/action.dart';
+import 'package:fl_clash/providers/app.dart';
 import 'package:fl_clash/providers/config.dart';
 import 'package:fl_clash/state.dart';
 import 'package:fl_clash/widgets/widgets.dart';
@@ -29,11 +30,13 @@ class DeveloperView extends ConsumerWidget {
           minVerticalPadding: 12,
           onTap: () {
             for (int i = 0; i < 1000; i++) {
-              appController.addLog(
-                Log.app(
-                  '[$i]${utils.generateRandomString(maxLength: 200, minLength: 20)}',
-                ),
-              );
+              globalState.container
+                  .read(logsProvider.notifier)
+                  .add(
+                    Log.app(
+                      '[$i]${utils.generateRandomString(maxLength: 200, minLength: 20)}',
+                    ),
+                  );
             }
           },
         ),
@@ -61,7 +64,9 @@ class DeveloperView extends ConsumerWidget {
             if (res != true) {
               return;
             }
-            await appController.handleClear();
+            await globalState.container
+                .read(storeActionProvider.notifier)
+                .handleClear();
           },
         ),
         // ListItem(
@@ -76,8 +81,10 @@ class DeveloperView extends ConsumerWidget {
         ListItem(
           title: Text(appLocalizations.pruneCache),
           minVerticalPadding: 12,
-          onTap: () {
-            appController.shakingStore();
+          onTap: () async {
+            await globalState.container
+                .read(storeActionProvider.notifier)
+                .shakingStore();
           },
         ),
       ],
