+++ lib/views/profiles/edit.dart
@@ -3,11 +3,11 @@ import 'dart:io';
 import 'dart:typed_data';
 
 import 'package:fl_clash/common/common.dart';
-import 'package:fl_clash/controller.dart';
 import 'package:fl_clash/core/controller.dart';
 import 'package:fl_clash/enum/enum.dart';
 import 'package:fl_clash/models/models.dart';
 import 'package:fl_clash/pages/editor.dart';
+import 'package:fl_clash/providers/action.dart';
 import 'package:fl_clash/state.dart';
 import 'package:fl_clash/widgets/widgets.dart';
 import 'package:flutter/material.dart';
@@ -71,6 +71,9 @@ class _EditProfileViewState extends State<EditProfileView> {
         minutes: int.parse(_autoUpdateDurationController.text),
       ),
     );
+    final profilesAction = globalState.container.read(
+      profilesActionProvider.notifier,
+    );
     final hasUpdate = widget.profile.url != profile.url;
     if (_fileData != null) {
       if (profile.type == ProfileType.url && _autoUpdate) {
@@ -83,14 +86,14 @@ class _EditProfileViewState extends State<EditProfileView> {
           profile = profile.copyWith(autoUpdate: false);
         }
       }
-      appController.putProfile(await profile.saveFile(_fileData!));
+      profilesAction.putProfile(await profile.saveFile(_fileData!));
     } else if (!hasUpdate) {
-      appController.putProfile(profile);
+      profilesAction.putProfile(profile);
     } else {
-      appController.safeRun(() async {
+      globalState.safeRun(() async {
         await Future.delayed(commonDuration);
         if (hasUpdate) {
-          await appController.updateProfile(profile);
+          await profilesAction.updateProfile(profile);
         }
       });
     }
@@ -107,7 +110,7 @@ class _EditProfileViewState extends State<EditProfileView> {
   }
 
   Future<void> _handleSaveEdit(BuildContext context, String data) async {
-    final message = await appController.safeRun<String>(() async {
+    final message = await globalState.safeRun<String>(() async {
       final message = await coreController.validateConfigWithData(data);
       return message;
     }, silence: false);
@@ -172,7 +175,7 @@ class _EditProfileViewState extends State<EditProfileView> {
   }
 
   Future<void> _uploadProfileFile() async {
-    final platformFile = await appController.safeRun(picker.pickerFile);
+    final platformFile = await globalState.safeRun(picker.pickerFile);
     if (platformFile?.bytes == null) return;
     _fileData = platformFile?.bytes;
     if (!mounted) {
@@ -206,7 +209,7 @@ class _EditProfileViewState extends State<EditProfileView> {
     _fileInfoNotifier.dispose();
     _autoUpdateDurationController.dispose();
     super.dispose();
-    appController.autoApplyProfile();
+    globalState.container.read(setupActionProvider.notifier).autoApplyProfile();
   }
 
   @override
