+++ lib/views/proxies/card.dart
@@ -1,5 +1,4 @@
 import 'package:fl_clash/common/common.dart';
-import 'package:fl_clash/controller.dart';
 import 'package:fl_clash/enum/enum.dart';
 import 'package:fl_clash/models/models.dart';
 import 'package:fl_clash/providers/providers.dart';
@@ -99,14 +98,19 @@ class ProxyCard extends StatelessWidget {
   Future<void> _changeProxy(WidgetRef ref) async {
     final isComputedSelected = groupType.isComputedSelected;
     final isSelector = groupType == GroupType.Selector;
+    final ref = globalState.container;
     if (isComputedSelected || isSelector) {
-      final currentProxyName = ref.read(getProxyNameProvider(groupName));
+      final currentProxyName = ref.read(proxyNameProvider(groupName));
       final nextProxyName = switch (isComputedSelected) {
         true => currentProxyName == proxy.name ? '' : proxy.name,
         false => proxy.name,
       };
-      appController.updateCurrentSelectedMap(groupName, nextProxyName);
-      appController.changeProxyDebounce(groupName, nextProxyName);
+      ref
+          .read(profilesActionProvider.notifier)
+          .updateCurrentSelectedMap(groupName, nextProxyName);
+      ref
+          .read(proxiesActionProvider.notifier)
+          .changeProxyDebounce(groupName, nextProxyName);
       return;
     }
     globalState.showNotifier(currentAppLocalizations.notSelectedTip);
@@ -122,7 +126,7 @@ class ProxyCard extends StatelessWidget {
         Consumer(
           builder: (_, ref, child) {
             final selectedProxyName = ref.watch(
-              getSelectedProxyNameProvider(groupName),
+              selectedProxyNameProvider(groupName),
             );
             return CommonCard(
               key: key,
@@ -198,7 +202,7 @@ class _ProxyDesc extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final desc = ref.watch(getProxyDescProvider(proxy));
+    final desc = ref.watch(proxyDescProvider(proxy));
     return EmojiText(
       desc,
       overflow: TextOverflow.ellipsis,
@@ -217,7 +221,7 @@ class _ProxyComputedMark extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final proxyName = ref.watch(getProxyNameProvider(groupName));
+    final proxyName = ref.watch(proxyNameProvider(groupName));
     if (proxyName != proxy.name) {
       return SizedBox();
     }
